tests: organize contract tests in sub dirs
