--- test/acl_interpreter.js
@@ -1,3 +1,0 @@
-const runSolidityTest = require('./helpers/runSolidityTest')
-
-runSolidityTest('TestACLInterpreter')
