+++ test/contracts/acl/acl_interpreter.js
@@ -1,0 +1,3 @@
+const runSolidityTest = require('../../helpers/runSolidityTest')
+
+runSolidityTest('TestACLInterpreter')
