+++ test/contracts/common/conversion_helpers.js
@@ -1,0 +1,3 @@
+const runSolidityTest = require('../../helpers/runSolidityTest')
+
+runSolidityTest('TestConversionHelpers')
