+++ test/contracts/common/delegate_proxy.js
@@ -1,0 +1,3 @@
+const runSolidityTest = require('../../helpers/runSolidityTest')
+
+runSolidityTest('TestDelegateProxy')
