+++ test/contracts/common/reentrancy_guard.js
@@ -1,4 +1,4 @@
-const { assertRevert } = require('./helpers/assertThrow')
+const { assertRevert } = require('../../helpers/assertThrow')
 
 const ZERO_ADDR = '0x0000000000000000000000000000000000000000'
