+++ test/contracts/kernel/kernel_funds.js
@@ -1,6 +1,6 @@
-const { assertRevert } = require('./helpers/assertThrow')
-const { getBalance } = require('./helpers/web3')
-const { onlyIf } = require('./helpers/onlyIf')
+const { assertRevert } = require('../../helpers/assertThrow')
+const { getBalance } = require('../../helpers/web3')
+const { onlyIf } = require('../../helpers/onlyIf')
 
 const ACL = artifacts.require('ACL')
 const Kernel = artifacts.require('Kernel')
