--- test/delegate_proxy.js
@@ -1,3 +1,0 @@
-const runSolidityTest = require('./helpers/runSolidityTest')
-
-runSolidityTest('TestDelegateProxy')
